Revert "Attacks can now have knockback which will push the other player entities by a velocity determinewd by the attack and their current "damage". This is all still very POC. I need to bring move data out of the game and into external descriptors, and I need to spend a lot of time refining the knockback calculations."
This reverts commit f370f80063fb98c947bfe980bc9561753a036be5.
