+++ core/src/com/nighto/weebu/entity/Entity.java
@@ -108,8 +108,6 @@ public class Entity implements EventListener {
         yPrevPos = yPos;
         xPos = x;
         yPos = y;
-        xVel = 0;
-        yVel = 0;
     }
 
     public void teleport(float x, float y, boolean keepVelocity) {
