+++ core/src/com/nighto/weebu/entity/attack/Attack.java
@@ -6,6 +6,7 @@ import com.badlogic.gdx.math.Rectangle;
 import com.nighto.weebu.entity.Entity;
 import com.nighto.weebu.entity.character.Character;
 import com.nighto.weebu.event.events.Event;
+import com.nighto.weebu.screen.StageScreen;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -13,11 +14,7 @@ import java.util.stream.Collectors;
 public class Attack extends Entity {
     protected float startupTime;
     protected float attackTime;
-
     protected float knockbackInduced;
-    protected float knockbackModifierIncrease;
-    protected float xImpulse;
-    protected float yImpulse;
 
     private Character owner;
 
@@ -26,8 +23,6 @@ public class Attack extends Entity {
 
     protected Attack(Character owner, List<Rectangle> hitBoxes) {
         super(owner.getStageScreen());
-        setHidden(true);
-
         this.owner = owner;
         this.hitBoxes = hitBoxes;
 
@@ -74,18 +69,6 @@ public class Attack extends Entity {
         return getHitBoxes();
     }
 
-    @Override
-    public void spawn(float x, float y) {
-        // Keep velocity across spawns for attacks (this is a temp fix for lasers until I can refactor this)
-        float xVel = getxVel();
-        float yVel = getyVel();
-
-        super.spawn(x, y);
-
-        setxVel(xVel);
-        setyVel(yVel);
-    }
-
     public boolean isPlaying() {
         return startupTime > 0 || attackTime > 0;
     }
@@ -93,25 +76,5 @@ public class Attack extends Entity {
     public List<Rectangle> getHitBoxes() {
         // Translated by position
         return hitBoxes.stream().map(r -> new Rectangle(r.x + xPos, r.y + yPos, r.width, r.height)).collect(Collectors.toList());
-    }
-
-    public float getKnockbackInduced() {
-        return knockbackInduced;
-    }
-
-    public float getKnockbackModifierIncrease() {
-        return knockbackModifierIncrease;
-    }
-
-    public float getxImpulse() {
-        return xImpulse;
-    }
-
-    public float getyImpulse() {
-        return yImpulse;
-    }
-
-    public Entity getOwner() {
-        return owner;
     }
 }
