+++ core/src/com/nighto/weebu/entity/attack/MeleeAttack.java
@@ -7,16 +7,12 @@ import com.nighto.weebu.entity.character.Character;
 import java.util.Collections;
 
 public class MeleeAttack extends Attack {
-    public MeleeAttack(Character owner, float hitBoxOffsetX, float hitBoxOffsetY, float knockbackInduced, float xImpulse, float yImpulse) {
+    public MeleeAttack(Character owner, float hitBoxOffsetX, float hitBoxOffsetY) {
         super(owner, Collections.singletonList(new Rectangle(hitBoxOffsetX, hitBoxOffsetY, 20, 20)));
 
         startupTime = 1f/60f;
         attackTime = 3f/60f;
-
-        this.knockbackInduced = knockbackInduced;
-        this.xImpulse = xImpulse;
-        this.yImpulse = yImpulse;
-        this.knockbackModifierIncrease = 15;
+        knockbackInduced = 15f/60f;
     }
 
     @Override
