+++ core/src/com/nighto/weebu/entity/character/CharacterTimers.java
@@ -54,8 +54,4 @@ public class CharacterTimers {
     public float getKnockbackTimeRemaining() {
         return knockbackTimeRemaining;
     }
-
-    public void setKnockbackTimeRemaining(float knockbackTimeRemaining) {
-        this.knockbackTimeRemaining = knockbackTimeRemaining;
-    }
 }
