+++ core/src/com/nighto/weebu/entity/character/event/CollisionEventHandler.java
@@ -1,5 +1,6 @@
 package com.nighto.weebu.entity.character.event;
 
+import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.math.Rectangle;
 import com.badlogic.gdx.math.Shape2D;
 import com.nighto.weebu.controller.GameInput;
@@ -96,12 +97,7 @@ public class CollisionEventHandler implements EventHandler {
         }
 
         if (them instanceof Attack) {
-            Attack attack = (Attack) them;
-            Character character = (Character) us;
-
-            if (!attack.getOwner().equals(character)) {
-                character.enterKnockback(attack);
-            }
+            Gdx.app.log("Collision", "Collided with attack.");
         }
     }
 }
