+++ core/src/com/nighto/weebu/entity/player/input/handlers/AirborneStateInputHandler.java
@@ -14,8 +14,7 @@ public class AirborneStateInputHandler extends StateInputHandler {
     public AirborneStateInputHandler(Player player) {
         super(
                 player,
-                new State[]{State.AIRBORNE},
-                new State[]{State.SUBSTATE_KNOCKBACK}
+                new State[]{State.AIRBORNE}
         );
     }
